requirements.txt
# grid
python-binance
tinydb
python-telegram-handler

# web
flask
pandas

# dev
python-dotenv
